feat: fix testcase node
